Start some basic tests
